requirements.txt
alembic==0.9.9
attrs==18.1.0
click==6.7
Flask==1.0.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.1
marshmallow-sqlalchemy==0.13.2
more-itertools==4.1.0
mysqlclient==1.3.12
pluggy==0.6.0
py==1.5.3
pytest==3.5.1
python-dateutil==2.7.2
python-dotenv==0.8.2
python-editor==1.0.3
-e git+git@github.com:thomasweng15/robopubs.git@ef62067fdd8217909d998e1c16d20800f715a695#egg=robopubs
six==1.11.0
SQLAlchemy==1.2.7
uWSGI==2.0.17
Werkzeug==0.14.1
